# services/svc-payment/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Smart Parking - Payment Service API",
        "description": "OpenAPI docs cho svc-payment",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8002/api",
            "description": "Local API through svc-auth"
        }
    ],
    "paths": {
...
        "/peak-hours": {
            "get": {
                "tags": [
                    "Peak Hours"
                ],
                "summary": "Lấy danh sách giờ cao điểm",
                "description": "Lấy danh sách tất cả giờ cao điểm theo bãi đỗ xe",
                "operationId": "ceeffe194519b8b844c0b05fec7e29ac",
                "parameters": [
                    {
                        "name": "parking_lot_id",
                        "in": "query",
                        "description": "ID bãi đỗ xe",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "day_of_week",
                        "in": "query",
                        "description": "Ngày trong tuần (1-7, 1=Thứ 2)",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "maximum": 7,
                            "minimum": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Danh sách giờ cao điểm",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "parking_lot_id": {
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "day_of_week": {
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "start_time": {
                                                "type": "string",
                                                "example": "07:00:00"
                                            },
                                            "end_time": {
                                                "type": "string",
                                                "example": "09:00:00"
                                            },
                                            "is_active": {
                                                "type": "boolean",
                                                "example": true
                                            },
                                            "created_at": {
                                                "type": "string",
                                                "format": "datetime"
                                            },
                                            "updated_at": {
                                                "type": "string",
                                                "format": "datetime"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Peak Hours"
                ],
                "summary": "Tạo giờ cao điểm mới",
                "description": "Tạo một giờ cao điểm mới cho bãi đỗ xe",
                "operationId": "37021049a2a0d309d2d167e1a54ca13b",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "parking_lot_id",
                                    "day_of_week",
                                    "start_time",
                                    "end_time"
                                ],
                                "properties": {
                                    "parking_lot_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "day_of_week": {
                                        "description": "1=Thứ 2, 2=Thứ 3, ..., 7=Chủ nhật",
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "start_time": {
                                        "type": "string",
                                        "example": "07:00:00"
                                    },
                                    "end_time": {
                                        "type": "string",
                                        "example": "09:00:00"
                                    },
                                    "is_active": {
                                        "type": "boolean",
                                        "example": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Tạo giờ cao điểm thành công",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Tạo giờ cao điểm thành công"
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Dữ liệu không hợp lệ"
                    }
                }
            }
        },
        "/peak-hours/{id}": {
            "get": {
                "tags": [
                    "Peak Hours"
                ],
                "summary": "Lấy chi tiết giờ cao điểm",
                "description": "Lấy thông tin chi tiết của một giờ cao điểm",
                "operationId": "75ea0399158d058a8ce2ec252421a05b",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID giờ cao điểm",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Chi tiết giờ cao điểm",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "parking_lot_id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "day_of_week": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "start_time": {
                                                    "type": "string",
                                                    "example": "07:00:00"
                                                },
                                                "end_time": {
                                                    "type": "string",
                                                    "example": "09:00:00"
                                                },
                                                "is_active": {
                                                    "type": "boolean",
                                                    "example": true
                                                },
                                                "parking_lot": {
                                                    "type": "object"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Không tìm thấy giờ cao điểm"
                    }
                }
            },
            "put": {
                "tags": [
                    "Peak Hours"
                ],
                "summary": "Cập nhật giờ cao điểm",
                "description": "Cập nhật thông tin của một giờ cao điểm",
                "operationId": "22f7c4a12f283db81f6f96fe569e6b3f",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID giờ cao điểm",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "day_of_week": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "start_time": {
                                        "type": "string",
                                        "example": "08:00:00"
                                    },
                                    "end_time": {
                                        "type": "string",
                                        "example": "10:00:00"
                                    },
                                    "is_active": {
                                        "type": "boolean",
                                        "example": false
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Cập nhật giờ cao điểm thành công",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Cập nhật giờ cao điểm thành công"
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Không tìm thấy giờ cao điểm"
                    },
                    "422": {
                        "description": "Dữ liệu không hợp lệ"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Peak Hours"
                ],
                "summary": "Xóa giờ cao điểm",
                "description": "Xóa một giờ cao điểm",
                "operationId": "61841862a123b9766ad454e2bd1c30cc",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID giờ cao điểm",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Xóa giờ cao điểm thành công",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Xóa giờ cao điểm thành công"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Không tìm thấy giờ cao điểm"
                    }
                }
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}